Add files via upload
  ALARM CLOCK:
INTRODUCTION:  This project about alarm clock is to set a preferred time you want, once the preferred time is matched with the current time the alarm starts ringing.
EXPECTED  OUTPUT: When the programme is executed a tkinter dialog box of dimension (300*500) opens in which the time is set in terms of hour, minutes and seconds of 12 hour format. When it matches with the current time a dialog box with message ‘TIME UP ‘appears with alarm sound. 
STRUCTURE OF INPUT: 
•	Import modules: modules like Tkinter, datetime, winsound and time.
•	Module of actual time: actual time is set in terms of both 24 hour format and 12 hour format.
•	Module of setting the alarm: it includes alarmclock which checks whether clock time matches with alarm time.
•	While loop: it has basic conditions to run the program. 
•	Code for the GUI: root which includes main GUI window.
STRUCTURE OF OUTPUT:
•	A Tkinter dialog box pops out on the window with a dimension, font, colour, width, size and other basic characters when the programme is executed. 
•	The dialog box contains buttons and space to set up the preferred time needed.
•	When the time set up matches with current time then the alarm clock works with a beep sound and a message ‘time up’ pops up.
PYTHON MODULES AND OTHER TOOL:
•	TKINTER MODULE:  to create a dialog box with information to provide or get from the user.
•	DATETIME AND TIME MODULE: To work with current time and date of present day and to manipulate it.
•	WINSOUND MODULE: generate the sound immediately when tune is called.
•	GUI LIBRARY: For rendering graphics on a display window.